refactor: split thunk functions, types, and extraReducers from slice files
