+++ client/src/slices/room/slice.ts
@@ -1,15 +1,8 @@
-import { createAsyncThunk, createSlice, PayloadAction } from '@reduxjs/toolkit';
-import { RootState } from 'app/store';
-import { ApiResponse, CreateRoomDto, Room, Message, Chat } from 'digiroom-types';
-import { axios } from 'utils/axios.util';
-
-export interface RoomState {
-  username: string;
-  userId: string;
-  room: Room | null;
-  messages: Message[];
-  state: 'disconnected' | 'failed' | 'joining' | 'joined';
-}
+import { createSlice, PayloadAction } from '@reduxjs/toolkit';
+import { Room, Message } from 'digiroom-types';
+import { extraReducers } from './reducer';
+import { createRoom, getRoom, getChat, addUserToRoom, joinRoom } from './thunk';
+import { RoomState } from './types';
 
 const initialState: RoomState = {
   username: localStorage.getItem('username') ?? '',
@@ -47,143 +40,8 @@ export const roomSlice = createSlice({
       state.messages = initialState.messages;
     },
   },
-  // The `extraReducers` field lets the slice handle actions defined elsewhere,
-  // including actions generated by createAsyncThunk or in other slices.
-  extraReducers: builder => {
-    builder.addCase(createRoom.fulfilled, (state, action) => {
-      state.state = 'disconnected';
-      state.room = action.payload.data;
-      state.userId = action.payload.data.users[0]._id; // bc its first user its ok
-    });
-
-    builder.addCase(getRoom.fulfilled, (state, action) => {
-      state.room = action.payload.data;
-    });
-
-    builder.addCase(getChat.fulfilled, (state, action) => {
-      state.messages = action.payload.data.messages;
-    });
-
-    builder
-      .addCase(addUserToRoom.pending, (state, action) => {
-        state.username = action.meta.arg.username;
-      })
-      .addCase(addUserToRoom.fulfilled, (state, action) => {
-        const username = action.meta.arg.username;
-        state.userId =
-          action.payload.data.users.find(user => user.name === username)?._id ?? '';
-        state.room = action.payload.data;
-      });
-
-    builder
-      .addCase(joinRoom.pending, state => {
-        state.state = 'joining';
-      })
-      .addCase(joinRoom.fulfilled, (state, action) => {
-        state.state = 'joined';
-      });
-  },
+  extraReducers: builder => extraReducers(builder),
 });
-
-const createRoom = createAsyncThunk(
-  `${roomSlice.name}/createRoom`,
-  async (payload: Omit<CreateRoomDto, 'username'>, { rejectWithValue, getState }) => {
-    try {
-      const username = (getState() as RootState).room.username;
-      const { data } = await axios.post<ApiResponse<Room>>('/room', {
-        ...payload,
-        username,
-      });
-
-      const userId = data.data.users[0]._id;
-      localStorage.setItem('userId', userId);
-
-      return data;
-    } catch (e) {
-      return rejectWithValue(e);
-    }
-  },
-);
-
-const getRoom = createAsyncThunk(
-  `${roomSlice.name}/getRoom`,
-  async (roomName: string, { rejectWithValue }) => {
-    try {
-      const { data } = await axios.get<ApiResponse<Room>>(`/room/${roomName}`);
-
-      return data;
-    } catch (e) {
-      return rejectWithValue(e);
-    }
-  },
-);
-
-const getChat = createAsyncThunk(
-  `${roomSlice.name}/getChat`,
-  async (roomName: string, { rejectWithValue }) => {
-    try {
-      const { data } = await axios.get<ApiResponse<Chat>>(`/room/${roomName}/chat`);
-
-      return data;
-    } catch (e) {
-      return rejectWithValue(e);
-    }
-  },
-);
-
-const addUserToRoom = createAsyncThunk(
-  `${roomSlice.name}/addUserToRoom`,
-  async (
-    { roomName, username }: { roomName: string; username: string },
-    { rejectWithValue },
-  ) => {
-    try {
-      const { data } = await axios.post<ApiResponse<Room>>(`/room/${roomName}/user`, {
-        username,
-      });
-
-      const userId = data.data.users.find(user => user.name === username)?._id ?? '';
-
-      localStorage.setItem('userId', userId);
-
-      return data;
-    } catch (e) {
-      return rejectWithValue(e);
-    }
-  },
-);
-
-const joinRoom = createAsyncThunk(
-  `${roomSlice.name}/joinRoom`,
-  async (
-    { roomName, callback }: { roomName: string; callback: (username: string) => void },
-    { rejectWithValue, dispatch, getState },
-  ) => {
-    try {
-      const username = (getState() as RootState).room.username;
-
-      let roomData = await dispatch(getRoom(roomName)).unwrap();
-
-      const isUsernameMember = roomData.data.users.find(user => user.name === username);
-
-      if (!isUsernameMember) {
-        const newRoomData = await dispatch(
-          addUserToRoom({ roomName, username }),
-        ).unwrap();
-
-        roomData = newRoomData;
-      }
-
-      callback(username);
-
-      const userId = roomData.data.users.find(user => user.name === username)?._id ?? '';
-
-      localStorage.setItem('userId', userId);
-    } catch (e) {
-      return rejectWithValue(e);
-    }
-  },
-);
 
 export const roomActions = {
   ...roomSlice.actions,
