+++ client/src/slices/room/types.ts
@@ -1,0 +1,9 @@
+import { Room, Message } from 'digiroom-types';
+
+export interface RoomState {
+  username: string;
+  userId: string;
+  room: Room | null;
+  messages: Message[];
+  state: 'disconnected' | 'failed' | 'joining' | 'joined';
+}
