+++ client/src/slices/search/types.ts
@@ -1,0 +1,8 @@
+import { YoutubePlaylistSearch, YoutubeVideoSearch } from 'youtube.ts';
+
+export interface SearchState {
+  youtubeSearchResult: YoutubeVideoSearch | YoutubePlaylistSearch | null;
+  state: 'idle' | 'loading' | 'failed';
+}
+
+export type YoutubeSearchForm = { searchTerm: string; type: 'playlists' | 'videos' };
