+++ client/src/slices/snackbar/slice.ts
@@ -1,9 +1,6 @@
-import { PayloadAction } from '@reduxjs/toolkit';
-import { createSlice } from '@reduxjs/toolkit';
-import { UnknownAsyncThunkRejectedAction } from '@reduxjs/toolkit/dist/matchers';
-import { AxiosError } from 'axios';
-import { ApiError } from 'digiroom-types';
+import { createSlice, PayloadAction } from '@reduxjs/toolkit';
 import { SnackbarKey } from 'notistack';
+import { extraReducers } from './reducer';
 import { ContainerState, SnackObjectAction } from './types';
 
 // The initial state of the Snackbar container
@@ -31,38 +28,7 @@ const snackbarSlice = createSlice({
       if (state.snacks?.[action.payload]) delete state.snacks[action.payload];
     },
   },
-  extraReducers: builder => {
-    builder.addMatcher(
-      action => action.type.endsWith('/rejected'),
-      (state, action: UnknownAsyncThunkRejectedAction) => {
-        const error = action.payload as AxiosError<ApiError>;
-        const defaultError = {
-          data: { message: 'Network problem probably, if not, contact dev' },
-        };
-        if (error.isAxiosError) {
-          const message = (error.response ?? defaultError).data.message;
-          state.snacks[message] = {
-            dismissed: false,
-            message: message,
-            key: message,
-            options: {
-              variant: 'error',
-            },
-          };
-        } else {
-          const key = error.message;
-          state.snacks[key] = {
-            dismissed: false,
-            message: 'contact the developer, error: ' + error.message,
-            key: key,
-            options: {
-              variant: 'error',
-            },
-          };
-        }
-      },
-    );
-  },
+  extraReducers: builder => extraReducers(builder),
 });
 
 export const snackbarActions = { ...snackbarSlice.actions };
